+++ packages/components/src/components/feedback/toast/toast/toast-styles.ts
@@ -1,7 +1,7 @@
 import {
   addImportantToEachRule,
   buildHostStyles,
-  contentWrapperMargin,
+x^  contentWrapperVars,
   getCss,
   mediaQuery,
   pxToRemWithUnit,
@@ -10,27 +10,29 @@ import { TOAST_Z_INDEX } from '../../../../constants';
 import { getAnimationIn, getAnimationOut, getKeyframesMobile } from '../../banner/banner-styles';
 import { TOAST_ANIMATION_DURATION_VAR } from './toast-manager';
 
-const toastPositionBottomVar = '--p-toast-position-bottom';
-const toastPositionBottomVarInternal = '--p-toast-position-bottom-internal';
+const toastPositionBottomVarPublic = '--p-toast-position-bottom'; // CSS custom property exposed as public interface
+const toastPositionBottomVarInternal = `${toastPositionBottomVarPublic}-internal`;
 export const toastCloseClassName = 'close';
+
+const { margin } = contentWrapperVars;
 
 export const getComponentCss = (): string => {
   return getCss({
     ...buildHostStyles(
       addImportantToEachRule({
         position: 'fixed',
-        left: contentWrapperMargin,
-        right: contentWrapperMargin,
+        left: margin,
+        right: margin,
         // Needs a not overwritable internal css variable to cover default position depending on viewport size and to handle animation properly.
         // In addition, a public css variable can be passed to overwrite the default position.
-        [toastPositionBottomVarInternal]: `var(${toastPositionBottomVar}, ${pxToRemWithUnit(56)})`,
+        [toastPositionBottomVarInternal]: `var(${toastPositionBottomVarPublic}, ${pxToRemWithUnit(56)})`,
         bottom: `var(${toastPositionBottomVarInternal})`,
         maxWidth: '42rem',
         zIndex: TOAST_Z_INDEX,
         [mediaQuery('s')]: {
           left: pxToRemWithUnit(64),
           right: 'auto',
-          [toastPositionBottomVarInternal]: `var(${toastPositionBottomVar}, ${pxToRemWithUnit(64)})`,
+          [toastPositionBottomVarInternal]: `var(${toastPositionBottomVarPublic}, ${pxToRemWithUnit(64)})`,
           bottom: `var(${toastPositionBottomVarInternal})`,
         },
       })