+++ packages/components/CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### [Unreleased]
 
+### [2.7.0-rc.0] - 2021-12-14
+
+#### Removed
+- `offset-bottom` prop of `Toast` (use `--p-toast-position-bottom` CSS variable instead)
+
 ### [2.7.0-beta.6] - 2021-12-08
 
 #### Added